fixed deprecation.

+++ main/src/com/polites/android/VectorF.java
@@ -15,7 +15,6 @@
  */
 package com.polites.android;
 
-import android.annotation.SuppressLint;
 import android.graphics.PointF;
 import android.util.FloatMath;
 import android.view.MotionEvent;
@@ -27,10 +26,10 @@ public class VectorF {
 	
 	public final PointF start = new PointF();
 	public final PointF end = new PointF();
-
+	
 	public void calculateEndPoint() {
-		end.x = (float)(Math.cos(angle) * length + start.x);
-		end.y = (float)(Math.sin((float)angle) * length + start.y);
+		end.x = (float)Math.cos(angle) * length + start.x;
+		end.y = (float)Math.sin(angle) * length + start.y;
 	}
 	
 	public void setStart(PointF p) {
